ats-checker/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
pdfplumber==0.11.8
spacy==3.8.3
groq==0.31.1
python-dotenv==1.0.0
django-cors-headers==4.6.0
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
